pyproject.toml
[tool.poetry]
name = "sacar"
version = "0.0.1.alpha23"
description = ""
authors = ["Kolonial.no <tech@kolonial.no>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.9"
starlette = "^0.13.0"
hypercorn = "^0.9.0"
python-multipart = "^0.0.5"
httpx = "^0.11.0"
dacite = "^1.2.0"
typing_extensions = "^3.7.4"
pyjwt = "^1.7.1"
cryptography = "^2.8"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = "^4.3.21"
flake8 = "^3.7.9"
mypy = "^0.761"
flake8-bugbear = "^20.1.2"
IPython = "^7.11.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
